clean up all scripts before PR merge
